Refactor landing scripts into modules
